+++ apps/ai-blog-writer/apps/backend/app/features/youtube2blog/graph/nodes/__init__.py
@@ -1,0 +1,1 @@
+"""Cohesive node families for the YouTube2Blog graph."""
